requirements.txt
frosty-ai>=1.5.0